fix: cambiamos el atributo id endomains a long para concordar con entity

+++ src/main/java/com/mgcss/domain/Cliente.java
@@ -1,13 +1,13 @@
 package com.mgcss.domain;
 
 public class Cliente {
-	private int id;
+	private long id;
 	private String nombre;
 	private String email;
 	private TipoCliente tipo;
 	
 	
-	public Cliente(int id, String nombre, String email, TipoCliente tipo){
+	public Cliente(long id, String nombre, String email, TipoCliente tipo){
 		this.id = id;
 		this.nombre = nombre;
 		this.email = email;
@@ -15,12 +15,12 @@ public class Cliente {
 	}
 
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
