+++ src/main/java/com/mgcss/domain/Tecnico.java
@@ -2,23 +2,23 @@ package com.mgcss.domain;
 
 public class Tecnico {
 
-	private int id;
+	private long id;
 	private String nombre;
 	private EspecialidadTecnico especialidad;
 	private boolean activo;
 	
-	public Tecnico(int id, String nombre, EspecialidadTecnico especialidad, boolean activo) {
+	public Tecnico(long id, String nombre, EspecialidadTecnico especialidad, boolean activo) {
 		this.id = id;
 		this.nombre = nombre;
 		this.especialidad = especialidad;
 		this.activo = activo;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
